Rename labels JS files to kebab-case
