+++ src/controllers/company-foreign.controller.js
@@ -6,7 +6,7 @@ const companyRepository = require('../repos/company.repo')
 const companyFormService = require('../services/company-form.service')
 const companyService = require('../services/company.service')
 const companyFormattingService = require('../services/company-formatting.service')
-const { companyDetailsLabels, accountManagementDisplayLabels, hqLabels } = require('../labels/companylabels')
+const { companyDetailsLabels, accountManagementDisplayLabels, hqLabels } = require('../labels/company-labels')
 const metadataRepository = require('../repos/metadata.repo')
 const { genCSRF, containsFormData, isBlank } = require('../lib/controller-utils')
 const router = express.Router()
