+++ src/controllers/company-ltd.controller.js
@@ -7,7 +7,7 @@ const companyFormService = require('../services/company-form.service')
 const companyFormattingService = require('../services/company-formatting.service')
 const companyRepository = require('../repos/company.repo')
 const metadataRepository = require('../repos/metadata.repo')
-const { companyDetailsLabels, chDetailsLabels, accountManagementDisplayLabels, hqLabels } = require('../labels/companylabels')
+const { companyDetailsLabels, chDetailsLabels, accountManagementDisplayLabels, hqLabels } = require('../labels/company-labels')
 const { genCSRF, containsFormData, isBlank } = require('../lib/controller-utils')
 const router = express.Router()
 const companyWithCHKeys = ['alias', 'trading_address', 'uk_region', 'headquarter_type', 'sector', 'website', 'description', 'employee_range', 'turnover_range']
