+++ src/controllers/contact-edit.controller.js
@@ -3,7 +3,7 @@ const Q = require('q')
 const winston = require('winston')
 const { genCSRF, containsFormData } = require('../lib/controller-utils')
 const contactFormService = require('../services/contact-form.service')
-const { contactLabels } = require('../labels/contactlabels')
+const { contactLabels } = require('../labels/contact-labels')
 const metadataRepository = require('../repos/metadata.repo')
 const companyRepository = require('../repos/company.repo')
 const { getCommon } = require('../controllers/contact.controller')
