+++ src/controllers/interaction.controller.js
@@ -2,7 +2,7 @@
 const express = require('express')
 const winston = require('winston')
 const Q = require('q')
-const interactionLabels = require('../labels/interactionlabels')
+const interactionLabels = require('../labels/interaction-labels')
 const {genCSRF} = require('../lib/controller-utils')
 const metadataRepository = require('../repos/metadata.repo')
 const interactionDataService = require('../services/interaction-data.service')
