+++ src/labels/investment-labels.js
